+++ app/Http/Controllers/ProductController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Services\ActivityLogService;
 use App\Models\Product;
 use App\Http\Services\ProductService;
 use App\Models\Category;
@@ -41,7 +42,7 @@ class ProductController extends Controller
     /**
      * Store a newly created resource in storage.
      */
-    public function store(Request $request)
+    public function store(Request $request, ActivityLogService $logService)
     {
         $data = $request->validate([
             'name' => 'required|string',
@@ -56,6 +57,13 @@ class ProductController extends Controller
         ]);
 
         $product = $this->productService->createProduct($data);
+        $supplierName = Supplier::find($data['supplier_id'])->name;
+        $categoryName = Category::find($data['category_id'])->name;
+
+        $logService->log(
+            'create_transaction',
+            "Menambahkan product {$data['name']} , sku : {$data['sku']}, supplier :$supplierName, category : $categoryName"
+        );
 
         if($request->hasFile('image')){
             $file = $request->file('image');
@@ -63,6 +71,11 @@ class ProductController extends Controller
             $filename = $product->id. '_' . time() . '.' . $file->getClientOriginalExtension();
 
             $path = $file->storeAs('products',$filename,'public');
+
+            $logService->log(
+            'Add_image',
+            "Menambahkan image ke product {$data['name']}"
+        );
 
             $this->productService->updateProduct( $product->id,['image'=> $path]);
         }
@@ -96,7 +109,7 @@ class ProductController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, string $id)
+    public function update(Request $request, string $id, ActivityLogService $logService)
     {
         $product = $this->productService->getProductById($id);
 
@@ -130,16 +143,84 @@ class ProductController extends Controller
             $product->update(['image' => $path]);
         }
 
+        $old = $this->productService->getProductById($id);
+        $oldData = [
+            'name' => $old->name,
+            'sku' => $old->sku,
+            'category_id' => $old->category_id,
+            'supplier_id' => $old->supplier_id,
+            'description' => $old->description,
+            'purchase_price' => $old->purchase_price,
+            'selling_price' => $old->selling_price,
+            'image' => $old->image,
+            'minimum_stock' => $old->minimum_stock,
+        ];
+
         $this->productService->updateProduct($id, $data);
+
+        // Membandingkan field yang berubah
+        $changed = [];
+        foreach($data as $key => $newValue){
+            if (array_key_exists($key,$oldData) && $oldData[$key] != $newValue){
+                $changed[$key] = [
+                    'old' => $oldData[$key],
+                    'new' => $newValue
+                ];
+            }
+        }
+
+        // Membuat deskripsi data yang berubah
+        if (!empty($changed)){
+            $parts = [];
+            foreach ($changed as $field => $values){
+                // Menampilkan nama produk
+                if($field === 'product_id'){
+                    $oldName = optional($old->product)->name ?? '-';
+                    $newName = optional(Product::find($values['new']))->name ?? '-';
+                    $parts[] = "produk: {$oldName} → {$newName}";
+                } else {
+                    $parts[] = "{$field}: {$values['old']} → {$values['new']}}";
+                }
+            }
+
+            $description = "Mengubah product ID {$old->id} (" .
+                implode(', ', $parts) . ")";
+
+            $logService->log('edit_product', $description);
+        }
+
+        // Membuat deskripsi log
+        $description = sprintf(
+            'Mengubah product'
+        );
+
         return redirect()->route('product.index');
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(string $id)
+    public function destroy(string $id, ActivityLogService $logService)
     {
+        $product = $this->productService->getProductById($id);
+        if(!$product){
+            return redirect()
+            ->route('product.index')
+            ->with('error', 'Transaksi tidak ditemukan');
+        }
+
+        // hapus product
         $this->productService->deleteProduct($id);
+
+        // Buat log
+        $logService->log(
+            'delete_product',
+            "Menghapus transaksi product ID {$product->id}, ".
+            "Produk {$product->name}, ".
+            "SKU {$product->sku}, ".
+            "Supplier {$product->supplier}, ".
+            "Category {$product->category}, "
+        );
         return redirect()->route('product.index');
     }
 }
